+++ docs/ChangeLog.md
@@ -102,6 +102,7 @@
 - main: work around nRF Connect initialization (requested by liyafe1997) `#D2129` 2df3b109
 - main(unload): redirect streams to work around trap `EXIT` in bash-5.2 (reported by ragnarov) `#D2142` 38a8d571
 - complete: call the `docker` command through `ble/util/conditional-sync` `#D2150` xxxxxxxx
+- util(joblist): fix job detection in Bash 5.3 `#D2157` xxxxxxxx
 
 ## Contrib